+++ src/h3lib/lib/coordijk.c
@@ -211,6 +211,11 @@ void _ijkScale(CoordIJK *c, int factor) {
  * Normalizes ijk coordinates by setting the components to the smallest possible
  * values. Works in place.
  *
+ * This function does not protect against signed integer overflow. The caller
+ * must ensure that none of (i - j), (i - k), (j - i), (j - k), (k - i), (k - j)
+ * will overflow. This function may be changed in the future to make that check
+ * itself and return an error code.
+ *
  * @param c The ijk coordinates to normalize.
  */
 void _ijkNormalize(CoordIJK *c) {
@@ -527,13 +532,40 @@ void ijkToIj(const CoordIJK *ijk, CoordIJ *ij) {
  *
  * @param ij The input IJ coordinates
  * @param ijk The output IJK+ coordinates
+ * @returns E_SUCCESS on success, E_FAILED if signed integer overflow would have
+ * occurred.
  */
-void ijToIjk(const CoordIJ *ij, CoordIJK *ijk) {
+H3Error ijToIjk(const CoordIJ *ij, CoordIJK *ijk) {
     ijk->i = ij->i;
     ijk->j = ij->j;
     ijk->k = 0;
 
+    // Check for the possibility of overflow
+    int max, min;
+    if (ijk->i > ijk->j) {
+        max = ijk->i;
+        min = ijk->j;
+    } else {
+        max = ijk->j;
+        min = ijk->i;
+    }
+    if (min < 0) {
+        // Only if the min is less than 0 will the resulting number be larger
+        // than max. If min is positive, then max is also positive, and a
+        // positive signed integer minus another positive signed integer will
+        // not overflow.
+        if (max < INT32_MIN - min) {
+            // max - min would overflow
+            return E_FAILED;
+        }
+        if (min == INT32_MIN) {
+            // 0 - INT32_MIN would overflow
+            return E_FAILED;
+        }
+    }
+
     _ijkNormalize(ijk);
+    return E_SUCCESS;
 }
 
 /**