+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The public API of this library consists of the functions declared in file
 ## [Unreleased]
 ### Fixed
 - Fixed possible signed integer overflow in `h3NeighborRotations` (#707)
+- Fixed possible signed integer overflow in `localIjToCell` (#706)
 
 ## [4.0.1] - 2022-09-15
 ### Fixed